Reducing warnings for ATLJETMET-703 / ATLASG-1258
In general, I think it's just fine if we apply a systematic that the
tool doesn't know about - it should just go ahead and return the nominal
response.  This sort of warning means that instead of just looping
through all tools and turning on a systematic, we have to either
suppress all the warnings out of these tools (obviously not great), or
we have to keep a map of which systematics apply to which tool and only
use the tools in the map (pretty heavy for most users).  Happily, the
developers agree with me, so this has been done by them in svn, and I'm
just making the same changes in git here.


Former-commit-id: 9eb559addafa5b55469a9b7acf69df939e3f34fd