InDetMonitoringATN: Fixes to get ATN tests working.
Update for running in git build.  Install test scripts to be able to find
them during testing.  Don't assume tests are being run from within
the package's directory.

For running from RAW files, TrigDecisionTool needs to be configured
to use TrigConfigSvc, not xAODConfigTool.  The latter requires metadata
that is not present when reading RAW.



Former-commit-id: a0e013ff31b5ef26ad8261952ab228f83a354b30