Fix segfault in dqm_algorithms DivideReference
The result object was being deleted by the use of boost::shared_ptr::reset
which then caused a segfault when the result was being published.

In the first case there is no need to replace the object since we're just adding
the new histogram to the existing collection (we don't want to delete the collection).

In the second case we _do_ want to replace the object with a new TCollection,
thus we need to add its clone to the collection to avoid the collection
holding a pointer to a deleted object.


Former-commit-id: 5982fc6dba9538efb09e6295dc6344d4c3e75cce