Adding Chargino Decays in Geant4
The option to decay Charginos to a pi+/pi- and a Neutralino within Geant4
existed in 19.2.X.Y, but was lost during the configuration migration
for MC16. This set of changes adds it back and in addition adds the
ability to have the Chargino decay to an e+/e- and a Neutralino or a
mu+/mu- and a Neutralino, as requested in ATLASSIM-3401.


Former-commit-id: f05abf81494977c927073280fd5ed526fffd5639