Adding an option to disable systematics for error-free running
When running pileup reweighting file creation (NTUP_PRW), the use of
systematics creates errors.  Not a shock - we are making the PRW file,
so we don't have one to work from.  Adding an option to disable the
systematics when running the derivation (default: run systematics so
that the default behavior is unchanged), and flipping that switch in the
derivation production to avoid the errors.


Former-commit-id: f0a825f6297b33401359af3bd32f70e0fd37172a