AthenaKernel: Change Timeout class to be slot-specific.
Timeout was a global singleton, used to flag when an algorithm reached a time limit.
This is used by some trigger algorithms.

A global singleton doesn't work with MT, however.  Change it so that it's
slot-local.

For clients accessing it with the existing interface, this requires that the
global event context is valid.  If not, the event context must be passed
explicitly.  I think this only affects the event loop managers.


Former-commit-id: 0bf67a7806cbd3584988863637e699f831ca7d07