Update interface of runJobs in backward-compat way
Several options of runJobs have been marked as deprecated, but their
functionality is (sometimes approximately) kept intact for now. New options
are added to replace and extend functionality.

RunJobs is typically invoked by other commands in beamspotman. It is important
that the command line arguments encode everything needed to reproduce the job
because they are stored in the task database and can be manually executed if a
job has failed, for example.

The new options don't have short flags, to be more explicit, and they are more
explicit about how the input file given is to be interpreted.


Former-commit-id: a41d3b50847a69ef016fd240f67afcd6d9571339