Standardise the extraction of dsname from run + tag
Several commands need to get a dataset name from the run number and tag. This
change moves the logic to do so into a helper function and removes the
arbitrary differences that previously existed between the similar chunks of
code.

In each case, the actual file list is not needed, but it is built in strict
mode to catch cases like missing files. All that is needed is the directory
path and the dataset basename, which are used to construct the command line
arguments to runJobs which builds and uses its own file set.


Former-commit-id: 59c287e5936e8bebcbe7445f0522e03f3c6cca1e