Check iCpm validity (ATR-16812)
Details:
this bug was introduced in !1359. Although rarely, it can lead to
events crashing the HLT, see ATR-16812.

Given the special conditions under which these events were being
recorded (one processor being power-cycled), it's not clear to me how
the software should deal with this circumstance. I believe that the
iCpm index should always be valid when in running state.


Former-commit-id: 5e0bafb1b172a7c2e91a60bba20cd9eaaa4f67e3