Remove requirement for muons, electrons, photons to be present to run AOD-reduction G4 truth thinning
ATLASRECTS-3392

This changes the behaviour  of the AOD-reduction G4 truth thinning, such that it no longer requires reconstructed electrons, muons and photons (which are normally needed since truth particles associated with these objects needs to be kept). Previously the job would crash (FATAL) if any of ```Muons```, ```Electrons``` or ```Photons``` were missing, and this was causing issues with subsystem-specific tests where not all containers may be defined. The code now just throws a warning and then carries on (and obviously doesn't try preserve G4 truth associated with the reconstructed objects that can't be found). It should be noted that if these container names were ever changed and the config file was not updated to reflect this, the code would ignore it apart from the warning, and then the G4 truth for reconstructed objects would be erroneously dropped.


Former-commit-id: 8df1b559407a344ac6d17a0f54e3907fcb1acc3d