Implemented additions to the existing error handling mechanism of the Event Service.
With these changes AthenaMP will send a new set of error messages to the pilot
when running in the Event Service mode. These messages inform the pilot about
non-critical errors, which may occur during processing of a given event range
(e.g. the requested event range was not found in the input file). As a result
of such errors only a given range needs to be declared as failed, while the
AthenaMP job will keep running and process new event ranges.


Former-commit-id: f982a1350253720f7e67ab42a0e48d43510f25be