Major cleanup of truth tools and GenFilter Migration
GenFilterTool is common, so has been moved to DerivationFrameworkMCTruth
instead of DerivationFrameworkSUSY.  Other derivations (Exot, JetM,
SUSY) have been modified according to that change.  Adding some bits to
the common DerivationFrameworkMCTruth to allow this to be setup for all
derivations, centrally.  Major cleanup of TRUTH1 and TRUTH3 derivations
in order to just use the common code as much as possible.  Should allow
removal of TRUTH4 once we are happy.  Also automatically sets up truth
jets and MET if we are running on EVNT, otherwise carries on (might be
useful for future truth-based derivations).


Former-commit-id: 6a89e64d925c7cde4bed72072b1ef60ce28698ac