Add improved exception reporting in the RoiDescriptor range check
The RoiDescriptor construction checks the input parameters
to ensure that bad parameters are not being provided by some calling algorithm
If these parameters are bad, then an RoiDescriptor can not be constructed
so an exception is thrown.

These changes specialise the exception to an specific range_error
exception that returns the out of range value through the what() method

Therefore this requires that correct handling of exceptions to report the
what() value.

It provides additional information to aid the debugging of upstream
algorithm problems such as in ATR-16631

It does not close ATR-16631, since the underlying issue there is in the
parameters passed into the RoiDescriptor, and not in the RoiDescriptor
itself, which is still behaviing correctly for the bad input values that
are provided


Former-commit-id: e9507f96aa91b3e6e37548805cb5cf2ad0bde0cd