Size reduction of DESDM_TILEMU by thinning CaloCellContainer
The CaloCellContainer is the dominant source of	DESDM_TILEMU size.
In this patch, the container is thinned by selecting only TILE cells
(caloDDE()->getSubCalo() == 3).

The TrackParticleContainers are	added into the DESDM_TILEMU to
perform	muon quality selections	in analyses. The size increase
is very	small compared to the total size.


Former-commit-id: 6bd53811cd6ec421c83582707ccd4e2db3cd3088