HFOR Tool Harmonization
Removing the run 1 HFOR tool (Generators/HforTool), which should not be
used any more.  Making sure that the new HFOR tool (in
PMGOverlapRemovalTools) gives clear messages about its actions.  Adding
the HFORTools to AnalysisBase, AthAnalysis, AthDerivation projects.  The
first two are so that it can be used by analyses without a special
checkout.  The latter is so that it can run in the derivation framework,
as a part of the core MCTruth derivations, so that analyses won't need
to use it.  If my master plan goes perfectly, then we will be able to
remove it from the analysis releases at a later date when everyone has
the information for free in their derivations.


Former-commit-id: d3de1693f98ea8443a4eba21c54ba35e6ec2bd75