Fixing unit test
At the moment we don't have an environmental variable holding a test
file in AthDerivation like we do in the analysis releases.  To get this
to work in AthDerivation, I'm pointing it at the current ASG MC test
file, and will point out that it might be useful to have a centrally
defined file on the ticket.


Former-commit-id: 8aee41607e99d2e0b3ae03a15c209911ccd72507