Set Default value of SimFlags.G4Commands
When we moved away from using FADS to set the G4 verbosity level we
switched to using a `SimFlag` to set a property of `G4AtlasAlg` and
`G4TransportTool` which then runs the command strings supplied via
`G4UIManager` interface. Although the configuration code attempted
to set the `SimFlag`, it was already locked and so nothing was set.
By adding the verbosity setting command as the default value of the
`SimFlag` it should always be supplied to `G4AtlasAlg` and
`G4TransportTool`.


Former-commit-id: ca25826f702d394fcd3d2cfcff90f3e7f1b8befc