AthenaServices: Make sure context is set before firing incidents.
In Hive, the (thread-local) current event context gets set by the scheduler,
before running an algorithm.  When an incident is fired, the thread-local
current context is retrieved and stored in the incident.  However, begin
and end run incidents are fired from AthenaHiveEventLoopMgr directly,
before the scheduler start processing the event.  Thus, these incidents
were getting a stale version of the event context.   Be sure that the
current context is set before firing incidents.


Former-commit-id: 255e1ea7ce2d5a8a3419ce27777eacdf8ef5f152