Converted all PixelMonModule* class instances to unique pointers
This makes sure that all allocated memory is properly taken care off
when the program finishes. In order not to run into segmentation
faults (due to double deletion attempts), the implemented destructors
of the PixelMonModule* classes, which call LWHist::safeDelete(hist)
for all histogram objects, were removed. The memory of those
histograms is taken care of by the AthenaMonitoring classes, where the
histograms are registered.


Former-commit-id: 1637cecd76e52327d9ebc4bce2fc7faf0383b607