Patches to truth setup
SUSYTruth and TauTruth truth matching configurations need to be explicit
with the MCTruthClassifier that they request.  This is a good reminder
that MCTruthClassifier should really be a service, not a tool, and that
will make our migration lives easier in the future.

Also adding quite a few additional checks in MCTruthCommon to try to
build and add to the sequence only what is needed in any given job.


Former-commit-id: ae82cf8e084cbd89d22b3e8b21c471ad14fed2bd