Cleaning up muon trigger SF tool configuration
Now that !4266 has been merged, we can significantly clean up the
SUSYTools muon trigger SF code.  There only needs to be a single tool,
and we can get rid of a lot of the old code switching tools, setting
years, and so on.  These tools now get the year from the `EventInfo`,
which also means the PRW tool is required (fine, that's already true for
egamma tools as well).


Former-commit-id: d6ea2d50b175f21b3cc3fd095aa85c653b32f86a