AthenaKernel: Remove EventIDBase comparison hacks.
ATLAS bases IOVs on run+lbn.  The new conditions code uses EventIDBase
from Gaudi to represent IOV times, but this did not do comparisons
properly in the run+lbn case.   A workaround was put in earlier
to deal with this.  The Gaudi class was recently updated to allow
the needed comparison, so now we can remove the workarounds added earlier.
A few other minor adjustments are also needed to get things working.



Former-commit-id: 2a65b11c634d592773d072e2ddcb8381be3f1869