Fixed missing Initialisation of covariance matrix (ATLASRECTS-4153).
The fast vertex fit without constraint passes a dummy vertex to the
fitter to indicate the starting point. The constraint is activated if
the trace of the covariance matrix is non zero. Prior to this fix the
covariance matrix was uninitialised, thus contained random values. As a
consequence the vertex fitter used a random constraint rather than no
constraint.


Former-commit-id: 8d40073c7fadcfe9112f35e7696ba9f3f321b075