Adapt PileUpTools to use the revised McEventCollection truth EDM.
The revised Truth EDM for RDO files writes one `McEventCollection`
for the signal event and then one `McEventCollection` for each
background type used in the digitization job. This is necessary
to remove the redundancy where HITS-level events with the same
event number could be used as hard-scatter and background. In such
a case all `HepMcParticleLinks` to that event number would point at
the hard-scatter truth as long as there was a particle with a matching
barcode in the hard-scatter event.

The `McEventCollectionHelper` is used to ensure that `HepMcParticleLink`
objects created by the `PileUpTools` point at the correct `McEventCollection`
when the new EDM is being used. This is not the default currently.

This pulls in changes from the following SVN tags:
BCM_Digitization-01-01-03-03
PixelDigitization-02-00-22-14
SCT_Digitization-02-00-16-06
TRT_Digitization-01-01-15-08
FastTRT_Digitization-00-02-07-03
sTGC_Digitization-00-04-07-02
TGC_Digitization-01-12-07-02
RPC_Digitization-01-05-17-02
MM_Digitization-00-05-23-02
MDT_Digitization-01-10-06-05
CSC_Digitization-01-15-01-02

It also applies similar updates to the `InnerDetector/InDetDigitization/FastSiDigitization`
package.

Additional changes are required to convert between the `xAOD` and
`PileUpTimeEventIndex` `PileUpType` enums in use in this branch
(see previous commit).

See ATLASSIM-2430.


Former-commit-id: 84dd2b4b568ae15517b6b51fdf556a002ea8baf7