AthenaServices: Make sure EventContext is set before firing incidents in AthenaMP.
In AthenaMP, we run beginRunAlgorithms as part of handling a BeforeFork
incident.  This can result in calls to IOVDbSvc, which require that a
valid EventContext object be available.  However,
AthenaEventLoopMgr::handle was not filling in the context.

Fix so that a valid context is provided.

Fixes ATN failures in AthenaMPTest.


Former-commit-id: 09d65399c9f27fac0ea433e6cc815e098591f802