Add HighPtMinBias to xAOD::EventInfo::PileUpType enum
This commit extends the `xAOD::EventInfo::PileUpType` by adding an
entry for `HighPtMinBias` bringing it into line with the enums
in `PileUpTimeEventIndex` and elsewhere. This update also adds a helper
function to convert the xAOD version of the enum into the
`PileUpTimeEventIndex` version, this will be needed until all of the
core digitization code has been updated to use xAOD types.
These changes are not present in the SVN version of the `HepMcParticleLink`
migration as the xAOD types are not used in the 20.20 branch.
This is part of the changes for ATLASSIM-2430.


Former-commit-id: 51859727f8ebd70bc7839711e05c578cc0e9b869