Fix for allAuthors not being present in some derivations
The new lowPt working point requires that allAuthors be present, but in many derivations it isn't.
This fix simply checks for the presence of allAuthors before applying the check that requires it.
This should allow those working on derivations that don't contain allAuthors to use this version until new derivations are available.


Former-commit-id: 3013466e8c9f00cf76fe856d35bdb94bea50a1a6