Cleaned-up `CaloHitAna`. Reduced `CaloHitAna` memory usage.
This is a manual sweep of the changes from !3978 into `master`.
Added an option to enable/disable recording of g4hits. The reason
for these changes is to fix the memory problems that were present
in the previous version of the code. There were minor problems with
the initialisation of new branches for the output tree, there were
some vectors that weren't getting cleared after each loop, and
there were a couple of objects that weren't being used anymore
which took up too much memory.

The initialisation of branches has been moved to the header file,
lines have been added to the source file to clean the vectors after
each loop and the unnecessary objects that accounted for most of the
memory problems have been removed.


Former-commit-id: 28f9dfe644439c35c228ddfa3bf9f33a875c377e