L1TopoSimulation: change UseBitwise default to True (ATR-14949)
Details:
the previous change bff74042 had switched the value of this parameter
for the TopoSteering instance that runs at HLT (which is an attribute
of TrigSteer), but not of the one in the monitoring (L1TopoSimulation
AthAlgorihtm).

Also add an `ATH_MSG_INFO` statement so that we know exactly what is
being run.

Please note that this switch has been already extensively tested in
ATR-15680.


Former-commit-id: 259c8affeef688d30b1a45ca35c1beb34ce8fc63