Adding a test for NTUP_PRW to the DF RTT
NTUP_PRW is a format that was previously untested, and as a result it
was left behind in the migration.  It was also broken by a merge request
recently, and we again were reminded how important having a test of some
kind is.  This simply adds a test for that format to the main derivation
framework RTT package.  The changes are from Monica Dobre, I'm just
checking them in here.


Former-commit-id: 855e358a563498cc29b7fa6dbb5d887049aa8bca