Add tool to decorate muons in DAODs with the covariance matrix of the associated InDetTrackParticle
ATLASG-1301

This MR adds a tools which decorate muons in DAODs with the covariance matrix of the associated InDetTrackParticle. This will enable us to drop the covariance matrix from the InDetTrackParticles smart slimming list, thereby saving a great deal of space in the DAODs. This change is mandated by the AOD task force.


Former-commit-id: b887f76b5108ae67a3156ad3880dc66ca4c2f6e3