Add command line option "--concurrent-events" for AthenaMT
Adds the command line option "--concurrent-events=N" which enables individual
control of the number of concurrent events as well as the number of threads in
an AthenaMT job.

If --concurrent-events OR --threads is used, the job is assumed to be MT.

If neither is set, then the job is not setup as AthenaMT.

If only one of --concurrent-events or --threads is used, the other takes the same
value (ie, same number of concurrent events as threads).

Both --concurrent-events and --threads can also be set simultaneously


Former-commit-id: f829830eb517bcf2516e136b1b90e96290285f4a