Removing obsolete files and updating some signal xsecs
GRLs are centrally kept on cvmfs now, so we don't need these ancient
GRLs in SUSYTools any more.  I'm also removing some of the ancient cross
section files, which shouldn't be needed for any R21 analysis.  Starting
with the ones that for sure aren't needed, and I'll follow up with folks
to see if we can do even more clean up.


Former-commit-id: 07429c2680d63fbe2d1948ff255a458a74e9bea8