Lower the priority of STACO muons
As the STACO algorithm is currently not	well tuned, it produces	mostly fakes.
This update lowers its priority	so that	muons (if any) reconstructed from the same ID candidate by the segment- or calo-tag algs are used instead.
In this	case, the MS track used by the STACO reconstruction will be reconstructed as a standalone muon,	so no information is lost.


Former-commit-id: 24d2720660d91d4977f4b63a94d92ceeda0bfebd