Moving OtherBSM to 46
After discussion with Attila and Christos, we agreed to keep the numbers
as specified for all already-existing definitions, and to add this new
one with the next number.  Of course, it doesn't matter where it is
written in the enum if we are being explicit, so I'm putting it where it
seems most reasonable in the list.


Former-commit-id: c4decd7c48c82995bab92195df0071190daf22bd