Making photon dressing consistency configurable
At the moment we have a photon dressing tool that adds photons to truth
leptons, and a truth jet building tool that tries to exclude photons
that are FSR off of a truth lepton.  There is no guarantee that these
are configured to get the same answer.  This is one step in that
direction.  With this change, the truth photon dressing tool can make a
decoration on the photons that says that they have been used in photon
dressing (according to it).  The truth jet building tool can be given
the same decoration to avoid picking up truth photons that were used in
dressing.

If a user sets both decorations to the same name, then photons will be
consistently ignored in both places.  This should be handled with care
at the moment -- it is important to remove some of the other options
(e.g. not also remove photons within dR<0.1 in the jet building code) in
order to get a fully consistent configuration.  This also forces an
order dependence -- if the decoration is desired, then photon dressing
must run before truth jet building (obviously).  Still, this allows us
physical consistency between the objects in a way that we have not had
up to now.

As a starting point, I think this should be used for muons and electrons,
as those are the obvious particles to dress.  It would be helpful to
discuss in a wider community whether it is sensible to also use this for
taus, even if the taus decay hadronically, and what ought to be done
with the truth jets in that situation.  This last issue relates to
ATLASG-930 and what ought to be done for those situations.  We will
raise this question with the PMG at some point in the future.

If a user does not set either decoration name, then the default behavior
should be unchanged, and there should be no additional CPU cost
associated with these updates.

While going through these changes I realized that the truth dressing
tool always runs DR dressing, and optionally runs anti-kT dressing as
well.  I think the behavior that we want is for it to run either anti-kT
or dR dressing, with dR dressing by default.  I've updated the tool at
the same time to use that construction.  An LXR search indicates that
there are no clients currently using anti-kT dressing, so we should be
safe.


Former-commit-id: b0a7a84ab7b435a2334d8655290585c6241f81a3