fix TAccept for multi-threaded use
The current pattern for using TAccept is to have it as a mutable class
member inside the selector tools and return a const reference from a
const-qualified accept() or select() method.  Needless to say that
this doesn't play well at all with AthenaMT coding rules.

To address this I split the existing class into two:
*  `TAcceptInfo`: contains all the data members that are set during
   initialize() and not touched afterwards.
*  `TAcceptData`: contains the actual bitmap that is generated when
    calling the selection tool.  This is now a lightweight object that
    can be returned by value, and which references back to the
    `TAcceptInfo` object for the shared data.

I kept the interface and implementation mostly as it was in the old
class, but I did switch from TString to std::string, used more of the
std::bitset functions where appropriate, and added some C++11
niceties.  I also marked the existing class and interfaces relying on
it as deprecated, as they are not MT-friendly.

Two implementation details I'm not sure about:
*  Right now I reference back to the TAcceptInfo object using a regular
   pointer, which means that the TAcceptData object shouldn't be kept
   around beyond the lifetime of the tool.  If we need it to stick
   around beyond that I should change it to a shared pointer now.
*  Looking at the client code it may be nice to have a TAcceptCut
   object that references a specific cut you configured and takes care
   of the boiler-plate error checking, etc.


Former-commit-id: 2e89e54f08ad0d06e815fb98088468513e45be01