adding BMG chambers to muon calibration stream
In order to get hits of a chamber written into the muon calibration
stream the chamber needs to be associated to the RoI and needs to
be decoded. As the BMG chambers are not equipped with RPCs they need
a special treatment. With this MR the chambers should end up in the stream.

@yyazawa confirmed he performed all test listed on
https://twiki.cern.ch/twiki/bin/viewauth/Atlas/TriggerDevelopmentGit
which are required for the MR. Furthermore, the request was discussed
with @pfleisch , @rosati , @sshaw , @mishitsu , @dellasta , @morgens .

Former-commit-id: a1c7490fcb42de0c6937af45c73091030e062313