Updating q/g label
After a discussion in the generator development meeting, we settled on a
slightly different standard for the 'common' truth flavor label.  If the
HF-tagging label is available, we will use it (as before).  If the
HF-tagging label is zero (unknown), we will use the Jet/MET label.  The
new bit is that if the Jet/MET label says that the jet is b/c/tau, and
the HF-tagging label said that it was unknown, then we will set the
label to 100 times the Jet/MET label in order to make this a clear
'something weird is going on' label.  This should not trick any of the
tools that are currently in use and look for b/c/tau/q/g labels.


Former-commit-id: f4e5b1fb8957b9a31353706ad360f1fa5be0cb18