Removing mutables and protecting against mis-configuration
Two mutable data members were being initialized during the execute.
Obviously this is thread-hostile, so we were going to have to clean it
up at some point anyway.  Now they are set during the init, which should
be safer in any case.

Since I introduced an init, also ensuring that the photon dressing
settings are consistent there.  In the case of inconsistent settings,
giving an error and returning failure, so that there's no way a user can
miss that their job is mis-configured.


Former-commit-id: 43eb5bbd40ea2fba6f631cc4bbe1757f52e22dc7