Improving OR option implementation, fixing UTs
The OR options as implemented before didn't quite do what I wanted them
to do -- rather than disabling the removal of electrons or muons, they
disabled both the removal of e/mu and the removal of jets caused by
e/mu.  This should fix that problem by being a little more specific
about what we want (note that it sets options to recover the e/mu in
both the case of sliding window "boosted" OR and in standard cone OR).

At the same time, fixing a stupid problem in the UTs -- the ilumicalc
files were specified separately, which meant that the second was
overwriting the first.


Former-commit-id: 4dc817331734c679a7398f13b5bcf443d9bf727c