Move Compressor class from EventCommonTPCnv to AthenaPoolCnvSvc
The `Compressor` class has no dependencies other than on the STL
and is not used by any other class in `EventCommonTPCnv` moving
it to `AthenaPoolCnvSvc` means that `CaloSimEventTPCnv` and
`LArSimEventTPCnv` no longer depend on `EventCommonTPCnv`.
In turn this means that `EventCommonTPCnv` no longer needs to be
part of the AthSimulation release. See ATLASSIM-3441.


Former-commit-id: e20bb7940afb3ce84074f560a8acb8fae0c3abc1