Gaudi v29: fixes for CondSvc and AlgExecStateSvc
This MR has some fundamental changes in how the condition
handling mechanism works.

  - CondInputLoader must declare all the base classes of the
objects  it will produce. This requires the use of SG::BaseInfoBase,
and the CondInputLoader must initialize after the libraries where
those bases are defined are loaded.

 - Condition objects must register themselves with a static
factory so that the CondInputLoader can create CondCont of the
correct type. This is done via the REGISTER_CC( classname ) macro
found in AthenaKernel/CondCont.h

 - ReadCondHandleKey will not create a CondCont if it isn't
present. If one is, they will cache a ptr to it. This means that
ReadCondHandle may have to do a store lookup for their CondCont
if the cached ptr is null.

There are some other minor changes:

 - AthenaHiveEventLoopMgr can show the timestamp in the event
   heartbeat if property ShowTimeStamp is set
 - a number of interfaces for the AlgExecStateSvc have changed
   in gaudi v29, and these are reflected in athena code


Former-commit-id: f6871bf30e1bbe6e577ac7e34640666af2dd8db3