LArCalibProcessing: Fix crashes in ATN tests.
The ATN tests have been crashing due to conditions access problems.
In particular, the jobs are configured for run 2 conditions access,
but the data files being processed are run 1.


Tried to rejigger things so that we don't crash.  However, i think the
conditions setup is probably still incorrect.

Actually, what should probably be done is to update these tests to run
on a run 2 data file.


Former-commit-id: d486b1d25f8af478bddb09dbae0271c2257d8529