EventSelectorAthenaPool: Change event selectors to use IEvtSelectorSeek instead of IEventSeek.
The event selector classes have used the IEventSeek interfaces to allow
for seeking in the input stream.  However, this is inconsistent with the
IEvtSelector interfaces.  The latter assume that the event selector is const
and put the state in a separate Context argument, which is passed to the
IEvtSelector methods.  (Never mind that none of the Atlas event selectors
are actually implemented like this.)  IEventSeek is used for the event
loop managers, which is fine, but the same interfaces are also used
for the event selectors.  These interfaces do not pass a Context argument,
resulting in the inconsistency.  This blocks any attempt to clean up
event selectors to remove the ugly const_casts and mutables with which
they have become adorned.

As a step towards improving this, introduce a new IEvtSelectorSeek interface,
which is a combination of the IEventSeek and ICollectionSize interfaces
where the methods are const and take a Context argument.

Changed all event selector classes to use this new interface, and changed
clients of the event selectors to use it.


Former-commit-id: 4e9dbcfa3d53d2e2a2cf83b3e3676aae7b0b5e2b