Significant protection to functions and ensuring passing arguments
Functionized truth tool set up as proposed by TJ in !4469 (I think that
one can be closed and we can continue to discuss over here, so that we
don't get collisions).  This lets us pass around parameters, like those
needed for keeping truth photon dressing decorations consistent.

Split the augmentation tools into two kernels, one of which does most of
the heavy lifting and runs before truth jet building, and one of which
does only the part that requires jets (and some of the other
collections) and which runs after truth jet building.

The functions that exist can now be called with an algSequence to which
everything should be added.  If they don't get an algSequence they
default to using the derivation master sequence.  This should also
address the performance penalty of sticking these things into the master
sequence too early, rather than keeping them in private sequencers that
can run after skimming has finished.

This *will* require client migration, but we want to get some discussion
of the code in first.


Former-commit-id: dc644552ec8b8d37b3bb92bb84c910a9d884417d