InDetAlignExample: Disable failing test.
The ATN test in this package has been failing for a long time because
the input file on which it depends was removed.
This was reported as ATLIDALIGN-7 months ago, but this report has been
ignored.  So, disable this test as there seems to be no interest
in actually fixing it.


Former-commit-id: 6a24768a5681dc9513d56baca2f25a306a2e98cc