Using G4 vertex information if available
Under some circumstances, in MC15 and MC16, we directly save the type of
Geant4 sub-process that created a particle.  Particularly in the case of
electrons, this can be used to easily distinguish a true electron from a
background electron, without having to navigate through the tree in any
complicated way.  Just adding a little code to take advantage of that.


Former-commit-id: 339f1a1c18bec2dabdea9f543107e3b94a9428e4