Updates to SUSYTools UT, failing on unrecognized option
Popping options out of rEnv as we use them allows us to check at the end
of the configuration whether there were any options set by the user that
were *not* consumed by the configuration.  This should help us catch
typos from the user or the use of obsolete options.  In the process,
caught a few cases of obsolete options being used (already working...)
and updated the unit tests accordingly.


Former-commit-id: 948e753c8701bb4ba1429bb17f3846494cef60ce