Python changes to get things running
Thanks to TJ's comments, I was able to get the jet algorithms up and
running again.  After this commit the last remaining issue is that the
decoration is not found on the particles, which I believe is because of
the const'ness of the truth particle collection in the truth dressing
tool.


Former-commit-id: 24222102131ccfd8151bdbd3f93b3ec2366b2a89