Python changes to get things running
Thanks to TJ's comments, I was able to get the jet algorithms up and
running again.  After this commit the last remaining issue is that the
decoration is not found on the particles, which I believe is because of
the const'ness of the truth particle collection in the truth dressing
tool.


Former-commit-id: 5c659af6c798176e51bd25b4ce5d4e02d2c224a5