Adding new decay collection maker
This is an intermediate commit.  In a few cases, the changes are all
white space (getting rid of tabs and newlines).  I've also added the
TruthDecayCollectionMaker, which is designed to eat a truth collection
and give back a new collection of particles and vertices that contain
all of the particles that have been specified along with all of their
decay products (and maintain navigability).


Former-commit-id: f8da2195e764039bcdab3389d2c4bc2114d6867f