Adding new decay collection maker
This is an intermediate commit.  In a few cases, the changes are all
white space (getting rid of tabs and newlines).  I've also added the
TruthDecayCollectionMaker, which is designed to eat a truth collection
and give back a new collection of particles and vertices that contain
all of the particles that have been specified along with all of their
decay products (and maintain navigability).


Former-commit-id: d74e3e4ad1821d98ac6bd70078b03e62adb3f1a3