Modified TrigRoiUpdater.cxx to use the == operator
Modified TrigRoiUpdater.cxx to use the == operator
when comparing RoIs instead of comparing coordinates
of each RoI. Operator can also catch cases of
duplicate composite RoIs.
Also fixed a typo in
DetectorDescrition/RoiDescriptor/Root/RoiUtil.cxx:
In the definition of the == operator when comparing
roi0.zedMinus() and roi1.zedMinus()
the if statement required the difference to be
>1e-7 for the two rois to be identical. Changed
it to <1e-7.


Former-commit-id: 8a5c91f93a92f20b803458bab1ed50ee5704c561