fix setActiveState value, and store flags in a TrigComposite
Details:
Werner explained to me that the active state is used by the steering
to determine whether the subsequent steps of the chain are executed.
In this case it means 'whether the event will be streamed' to the
debug stream.
Also store the internal flags to a TrigComposite object that can
be used to identify events from different sources (right now it's
not clear whether the events will go in our 'l1topo_debug' stream,
or in the l1calo one, or in the debug one).

Former-commit-id: 374ca3e6993dfbb685a44c4f34dc59cb144b2fbc