AthenaROOTAccess: Add return value to clearDecorations.
Add a return value to IConstAuxStore::clearDecorations() to tell if any
decorations were actually cleared.  Allows speeding up some implementations
of clearDecorations() for the case when there are actually no decorations.

This was breaking the hack we used to force the rebuild of the auxid
set in TBranchAlias::setStore when dynamic variables were copied
to a tp-converted store.  Rework that in a slightly cleaner manner
so that we don't need the hack any longer.



Former-commit-id: 7e851c2b9fc5202ae3180e126f5590176c725aff