Error code exit for DerivationFrameworkART jobs
This adds the ```set -e``` to all of the ART jobs, such that if either of the tasks in an ART job fail (DAOD building or merging) then the job returns a failure code and is marked red by ART. Without this MR only the merge job is used to detect the success or failure. The script-making script is also updated to reflect this modification.


Former-commit-id: db33bd38f72183a791dd13156dadab941f200f09