Update jet triggers in DAOD_JETM* (ATLJETMET-838)
The trigger list in DAOD_JETM* has not been updated since 2016
running.  This MR addresses the changes in jet triggers needed
for 2017 usage.  It also re-organized the list of triggers to
make it possible to read through the list and know what has or
has not been included.  Non-jet triggers are left to future MRs
hopefully with input from relevant trigger signature groups.


Former-commit-id: 5ed2c749b2c8c191bd70d2f880f4f0f83184f191