Fixed AthenaMP crash when running derivations with direct I/O (ATEAM-434)
Patched AthenaMPToolBase such that AthenaMP sub-processes will not attempt
to reopen file descriptor if either FileMgr or the FD Registry contains
open files with FD=-1. This value seems to be legitimately valid for
remote files.

The changes tested with RunTier0Tests on top of 21.0,Athena,r04.
The log can be found on AFS at:
/afs/cern.ch/user/t/tsulaia/public/RunTier0Tests-Oct6.log


Former-commit-id: 418eda1e2006e834e16441fe50db6a0e4de23fb8