SGComps: Fix TAD list handling in ProxyProviderSvc.
ProxyProviderSvc deals with lists of TransientAddress objects produced
by address providers.  An earlier set of changes reworked things a bit
so that a provider could edit the lists produced by earlier providers,
as a part of implmenting the renaming functionality.  However, this doesn't
work correctly if a provider removes items from the list, which we want
in order to implement having objects produced via a writehandle supercede
objects in the input file.

But in fact that earlier change was done in a needlessly complicated way.
Clean this up, so that it will also work in the face of deletions
from the TAD list.


Former-commit-id: b4a29548cc6bc090e3098d6795f3330f3bcc99d6