Fixing the removal of particles outside the max_eta range after the CS correction.
The particles with |eta|>max_eta were discarded after the CS correction
procedure. Now, they are kept unchanged in the output container.
Moreover, additional parameter is introduced which specifies the
rapidity range for the background estimation using
GridMedianBackgroundEstimator. The default value for the CS alpha
parameter is also changed - it corresponds to the optimal value
found in ATLAS-CONF-2017-065.


Former-commit-id: 30b5ca6551fbda0ab31759b7cbdce137c685c89f